% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.lifertableCIJackknife.R
\name{as.data.frame.lifertableCIJackknife}
\alias{as.data.frame.lifertableCIJackknife}
\title{Coerce CI to a Data Frame}
\usage{
\method{as.data.frame}{lifertableCIJackknife}(x, row.names = NULL, ...)
}
\arguments{
\item{x}{A lifertableCIJackknife object.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
\code{as.data.frame.lifertableCIJackknife} returns a data frame.
}
\description{
Function to coerce the object displaying the Confidence Interval into a data frame.
}
\examples{
## Make the main object:
lft <- lifertable(ColumnFemale = Female,
                  ColumnAge = Age,
                  ColumnEggs = Eggs,
                  SexRate = Sexrate,
                  ColumnGroups = Group,
                  data = Insects,
                  CI = TRUE)


as.data.frame(lft$CI)

}
