% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.lifertableLFT.R
\name{as.data.frame.lifertableLFT}
\alias{as.data.frame.lifertableLFT}
\title{Coerce Life Table to a Data Frame}
\usage{
\method{as.data.frame}{lifertableLFT}(x, row.names = NULL, ...)
}
\arguments{
\item{x}{A lifertableLFT object}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
\code{as.data.frame.lifertableLFT} returns a data frame.
}
\description{
Function to coerce the object displaying Life Table into a data frame.
}
\examples{
## Make the main object:
lft <- lifertable(Female, Age, Eggs, Sexrate,
                  ColumnGroups = Group, data = Insects)


as.data.frame(lft$LIFERTABLE)

}
