% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.lifertableParmEst.R
\name{as.data.frame.lifertableParmEst}
\alias{as.data.frame.lifertableParmEst}
\title{Coerce Parameters to a Data Frame}
\usage{
\method{as.data.frame}{lifertableParmEst}(x, row.names = NULL, ...)
}
\arguments{
\item{x}{A lifertableParmEst object}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
\code{as.data.frame.lifertableParmEst} returns a data frame.
}
\description{
Function to coerce the object displaying Parameters into a data frame.
}
\examples{
## Make the main object:
lft <- lifertable(Female, Age, Eggs, Sexrate,
                  ColumnGroups = Group, data = Insects)


as.data.frame(lft$PARAMETERS)

}
