% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.lifertableTotEggs.R
\name{as.data.frame.lifertableTotEggs}
\alias{as.data.frame.lifertableTotEggs}
\title{Coerce Total Eggs to a Data Frame}
\usage{
\method{as.data.frame}{lifertableTotEggs}(x, row.names = NULL, ...)
}
\arguments{
\item{x}{A lifertableTotEggs object.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
\code{as.data.frame.lifertableTotEggs} returns a data frame.
}
\description{
Function to coerce the object displaying Total Eggs into a data frame.
}
\examples{
## Make the main object:
lft <- lifertable(Female, Age, Eggs, Sexrate,
                  ColumnGroups = Group,
                  data = Insects, TotalEggs = TRUE)


as.data.frame(lft$TOTAL.EGGS)

}
