% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPseudoVals.R
\name{PlotPseudoVals}
\alias{PlotPseudoVals}
\title{Plot for the Pseudo-Values of Parameters}
\usage{
PlotPseudoVals(object)
}
\arguments{
\item{object}{An object inheriting from \code{\link[=lifertable]{lifertable()}}, representing the
Life and Fertility Table.}
}
\value{
Returns an object of \code{\link[base:class]{class}} c("gg", "ggplot").
}
\description{
Construction of Plots for the pseudo-values of associated Life Table parameters.
}
\examples{
## The main object will be created using the Insects database:
lft <- lifertable(Female, Age, Eggs, Sexrate, ColumnGroups = Group,
                  data = Insects, jackknife = TRUE)

PlotPseudoVals(lft)


}
